split out push_metrics into iota-common
