+++ crates/iota-common/Cargo.toml
@@ -7,8 +7,16 @@ license = "Apache-2.0"
 publish = false
 
 [dependencies]
+anyhow.workspace = true
+fastcrypto.workspace = true
 futures.workspace = true
+iota-metrics.workspace = true
+iota-tls.workspace = true
+iota-types.workspace = true
 parking_lot.workspace = true
+prometheus.workspace = true
+reqwest.workspace = true
+snap.workspace = true
 tokio = { workspace = true, features = ["sync"] }
 tracing.workspace = true
