+++ crates/iota-common/src/lib.rs
@@ -3,4 +3,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod logging;
+pub mod metrics;
 pub mod sync;
