+++ crates/iota-node/src/metrics.rs
@@ -2,53 +2,16 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use std::time::Duration;
 
-use axum::http::header;
+use iota_common::metrics::{MetricsPushClient, push_metrics};
 use iota_metrics::RegistryService;
 use iota_network::tonic::Code;
 use iota_network_stack::metrics::MetricsCallbackProvider;
 use prometheus::{
-    Encoder, HistogramVec, IntCounterVec, IntGaugeVec, PROTOBUF_FORMAT, Registry,
-    register_histogram_vec_with_registry, register_int_counter_vec_with_registry,
-    register_int_gauge_vec_with_registry,
+    HistogramVec, IntCounterVec, IntGaugeVec, Registry, register_histogram_vec_with_registry,
+    register_int_counter_vec_with_registry, register_int_gauge_vec_with_registry,
 };
-use tracing::error;
-
-const METRICS_PUSH_TIMEOUT: Duration = Duration::from_secs(45);
-
-pub struct MetricsPushClient {
-    certificate: std::sync::Arc<iota_tls::SelfSignedCertificate>,
-    client: reqwest::Client,
-}
-
-impl MetricsPushClient {
-    pub fn new(network_key: iota_types::crypto::NetworkKeyPair) -> Self {
-        use fastcrypto::traits::KeyPair;
-        let certificate = std::sync::Arc::new(iota_tls::SelfSignedCertificate::new(
-            network_key.private(),
-            iota_tls::IOTA_VALIDATOR_SERVER_NAME,
-        ));
-        let identity = certificate.reqwest_identity();
-        let client = reqwest::Client::builder()
-            .identity(identity)
-            .build()
-            .unwrap();
-
-        Self {
-            certificate,
-            client,
-        }
-    }
-
-    pub fn certificate(&self) -> &iota_tls::SelfSignedCertificate {
-        &self.certificate
-    }
-
-    pub fn client(&self) -> &reqwest::Client {
-        &self.client
-    }
-}
 
 /// Starts a task to periodically push metrics to a configured endpoint if a
 /// metrics push endpoint is configured.
@@ -76,63 +39,6 @@ pub fn start_metrics_push_task(config: &iota_config::NodeConfig, registry: Regis
     // metrics
     let config_copy = config.clone();
     let mut client = MetricsPushClient::new(config_copy.network_key_pair().copy());
-
-    async fn push_metrics(
-        client: &MetricsPushClient,
-        url: &reqwest::Url,
-        registry: &RegistryService,
-    ) -> Result<(), anyhow::Error> {
-        // now represents a collection timestamp for all of the metrics we send to the
-        // proxy
-        let now = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_millis() as i64;
-
-        let mut metric_families = registry.gather_all();
-        for mf in metric_families.iter_mut() {
-            for m in mf.mut_metric() {
-                m.set_timestamp_ms(now);
-            }
-        }
-
-        let mut buf: Vec<u8> = vec![];
-        let encoder = prometheus::ProtobufEncoder::new();
-        encoder.encode(&metric_families, &mut buf)?;
-
-        let mut s = snap::raw::Encoder::new();
-        let compressed = s.compress_vec(&buf).map_err(|err| {
-            error!("unable to snappy encode; {err}");
-            err
-        })?;
-
-        let response = client
-            .client()
-            .post(url.to_owned())
-            .header(reqwest::header::CONTENT_ENCODING, "snappy")
-            .header(header::CONTENT_TYPE, PROTOBUF_FORMAT)
-            .body(compressed)
-            .timeout(METRICS_PUSH_TIMEOUT)
-            .send()
-            .await?;
-
-        if !response.status().is_success() {
-            let status = response.status();
-            let body = match response.text().await {
-                Ok(body) => body,
-                Err(error) => format!("couldn't decode response body; {error}"),
-            };
-            return Err(anyhow::anyhow!(
-                "metrics push failed: [{}]:{}",
-                status,
-                body
-            ));
-        }
-
-        tracing::debug!("successfully pushed metrics to {url}");
-
-        Ok(())
-    }
 
     tokio::spawn(async move {
         tracing::info!(push_url =% url, interval =? interval, "Started Metrics Push Service");
